audio-inference/requirements.txt
flask
tensorflow-cpu
librosa
numpy
pydub
matplotlib